conformance: isolate generated Any registration
Build the large testset twice so ordinary conformance tests avoid cpp2:any static registration, whole-archive only the narrow registries that depend on startup side effects, and aggregate TestValue execution to avoid paying initializer cost across hundreds of processes.

+++ thrift/conformance/cpp2/internal/TestValue.h
@@ -103,7 +103,8 @@ uint32_t invoke(const WriteOp& write, W& writer) {
 
     default:
       folly::throw_exception<std::runtime_error>(
-          fmt::format("Unknown write call: {}", write.getType()));
+          fmt::format(
+              "Unknown write call: {}", static_cast<int>(write.getType())));
   }
 }
